Game loads images correctly, mario and objects move simultaneously

+++ src/main/java/com/parim/model/components/EnemyObject.java
@@ -2,7 +2,7 @@ package com.parim.model.components;
 
 public class EnemyObject extends TileObject {
     public EnemyObject(){}
-    public EnemyObject(int x, int y, String type) {
+    public EnemyObject(double x, double y, String type) {
         super(x, y, type);
     }
 }
